Fixed everything, ready to go

+++ LECTURE_04_ARP_POISON/topology.sh
@@ -8,7 +8,31 @@ echo "------------------------------------"
 echo "Autore: Patrizio Tufarolo <patrizio@tufarolo.eu>"
 echo
 
-Color_Off='\e[0m' 
+
+# MIT License
+#  
+# Copyright (c) 2018 Patrizio Tufarolo
+# 
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+# 
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+##
+
+Color_Off='\e[0m'
 bldgrn='\e[1;32m'
 
 ok_colorato_per_log() {
@@ -20,39 +44,39 @@ ok_colorato_per_log() {
         printf "%${COL}s%s%s%s%s%s" "$NORMAL" "[" "$GREEN" "OK" "$NORMAL" "]"
 }
 do_cmd() {
-	local cmd;
-	local msg;
-	cmd="$1"
-	msg="$2"
-	echo
-	echo -e $bldgrn $cmd $Color_Off;
-	eval "$cmd" 2>/dev/null
-	echo
-	echo -n "$msg"
-	ok_colorato_per_log ${#msg}
-	read -rsp $'\nPremere invio per continuare...\n'
+        local cmd;
+        local msg;
+        cmd="$1"
+        msg="$2"
+        echo
+        echo -e ${bldgrn}"${cmd}"${Color_Off}
+        eval "$cmd" 2>/dev/null
+        echo
+        echo -n "$msg"
+        ok_colorato_per_log ${#msg}
+        read -rsp $'\nPremere invio per continuare...\n'
 }
 do_cmd_pid() {
-	local cmd;
-	local msg;
-	local pid;
-	local __resultvar=$3;
-	cmd="$1"
-	msg="$2"
-	
-	echo
-	echo -e $bldgrn $cmd $Color_Off;
-	eval "$cmd & "
-	pid=$!
-	echo -n "$msg"
-	ok_colorato_per_log ${#msg}
-	read -rsp $'\nPremere invio per continuare...\n'
-	eval $__resultvar=$pid;
+        local cmd;
+        local msg;
+        local pid;
+        local __resultvar=$3;
+        cmd="$1"
+        msg="$2"
+
+        echo
+        echo -e $bldgrn $cmd $Color_Off;
+        eval "$cmd & "
+        pid=$!
+        echo -n "$msg"
+        ok_colorato_per_log ${#msg}
+        read -rsp $'\nPremere invio per continuare...\n'
+        eval $__resultvar=$pid;
 }
 
-
-
 sudo -i exit
+
+
 
 EXPERIMENT="$(sudo himage -e pc1)"
 
@@ -64,7 +88,15 @@ do_cmd "sudo himage pc1 ping -c 5 10.0.0.21"								"=> Lancio un ping da PC1 a 
 do_cmd "sudo himage pc1 ip neigh"									"=> Stampo la tabella ARP PC1"
 do_cmd "sudo himage pc2 ip neigh"									"=> Stampo la tabella ARP PC2"
 read -rsp $'\nPremere invio nuovamente per chiudere XTERM...\n'
-sudo kill -9 $pid1 1>/dev/null 2>/dev/null;
-sudo kill -9 $pid2 1>/dev/null 2>/dev/null;
-sudo kill -9 $pid3 1>/dev/null 2>/dev/null;
 
+if kill -0 $pid1; then 
+sudo kill -9 $pid1 1>/dev/null 2>/dev/null;
+fi
+
+if kill -0 $pid2; then
+sudo kill -9 $pid2 1>/dev/null 2>/dev/null;
+fi
+
+if kill -0 $pid3; then
+sudo kill -9 $pid3 1>/dev/null 2>/dev/null;
+fi
